+++ src/ast.rs
@@ -539,7 +539,7 @@ impl AsRef<Module> for AST {
 }
 
 /// An identifier containing a string name and a position.
-#[derive(Debug, Clone, Hash)]
+#[derive(Debug, Clone, Eq, PartialEq, Hash)]
 pub struct Ident {
     pub name: String,
     pub pos: Position,
@@ -553,7 +553,7 @@ impl Ident {
 }
 
 /// An identifier containing an immutable name and a position.
-#[derive(Debug, Clone, Hash)]
+#[derive(Debug, Clone, Eq, PartialEq, Hash)]
 pub struct IdentX {
     pub name: ImmutableString,
     pub pos: Position,
@@ -866,7 +866,7 @@ pub struct FnCallInfo {
     /// and the function names are predictable, so no need to allocate a new `String`.
     pub name: Cow<'static, str>,
     /// Namespace of the function, if any.
-    pub namespace: Option<Box<ModuleRef>>,
+    pub namespace: Option<ModuleRef>,
     /// Call native functions only? Set to `true` to skip searching for script-defined function overrides
     /// when it is certain that the function must be native (e.g. an operator).
     pub native_only: bool,
@@ -903,8 +903,8 @@ pub enum Expr {
     StringConstant(Box<IdentX>),
     /// FnPtr constant.
     FnPointer(Box<IdentX>),
-    /// Variable access - ((variable name, position), optional modules, hash, optional index)
-    Variable(Box<(Ident, Option<Box<ModuleRef>>, u64, Option<NonZeroUsize>)>),
+    /// Variable access - (variable name, optional modules, hash, optional index)
+    Variable(Box<(Ident, Option<ModuleRef>, u64, Option<NonZeroUsize>)>),
     /// Property access.
     Property(Box<(IdentX, (String, String))>),
     /// { stmt }